% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LineChart-functions.R
\name{applySettingsToPlottingDf}
\alias{applySettingsToPlottingDf}
\title{Apply appearance settings to a plotting data frame.}
\usage{
applySettingsToPlottingDf(settings, plotDf)
}
\arguments{
\item{settings}{The appearance settings to apply.}

\item{plotDf}{The plotting data frame to apply the settings to.}
}
\value{
A new plotting data frame with settings applied.
}
\description{
Apply appearance settings to a plotting data frame.
}
