% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LineChart-functions.R
\name{buildGroupSettings}
\alias{buildGroupSettings}
\title{Create a data frame with appearance settings for plotting}
\usage{
buildGroupSettings(
  group,
  groupLabel = NULL,
  color = NULL,
  fillColor = NULL,
  symbol = NULL,
  cex.symbol = NULL,
  width.errBar = NULL,
  lty = NULL,
  lwd = NULL,
  include = NULL,
  suppressWarnings = FALSE
)
}
\arguments{
\item{group}{The indentifiers for the groups.}

\item{groupLabel}{An alternative name to use in legends, instead of 'group'. If provided and some values are NA, those values will be replaced with the group name.}

\item{color}{The color to be used for the symbols and lines.}

\item{fillColor}{The fill color to be used in case the symbols are of a fillable type (e.g. 21:25).}

\item{symbol}{The plotting character to use.}

\item{cex.symbol}{The character expansion for the plotting symbols.}

\item{width.errBar}{The width of the heads of the error bars (if present).}

\item{lty}{The line type.}

\item{lwd}{The line width.}

\item{include}{Boolean. Should this group be included in plots?}

\item{suppressWarnings}{Boolean. By default you are warned about which settings are left at default values. Set to TRUE to suppress these warnings.}
}
\value{
A data frame containing appearance setting data that can be used with lineChart for the settings argument.
}
\description{
This function is the primary way to provide settings for the groups used in plotting.
You need to provide the indentifiers for each group and the various settings to be used
for that group. All settings are optional and will be set with reasonable defaults if
not specified.
}
\examples{
library(LineChart)

data(ChickWeight)
settings = buildGroupSettings(group=1:4, symbol=21:24, color=c("red", "green", "orange", "blue"))
lineChart( weight ~ Time * Diet, ChickWeight, settings = settings, legendPosition="topleft" )
}
