% Generated by roxygen2 (4.0.1): do not edit by hand
\name{createPlottingDf}
\alias{createPlottingDf}
\title{Creates a plottable data frame.}
\usage{
createPlottingDf(formula, data, settings = NULL, errBarType = "SE")
}
\arguments{
\item{formula}{The formula should be of the form y ~ x * group.}

\item{data}{A data frame containing the data to be plotted.}

\item{settings}{Plotting settings for the different groups in the data, such as the symbol to use.}

\item{errBarType}{String. The type of error bar to use. Can be "SE" for standard error, "SD" for standard deviation, or "CI95" for a 95\% confidence interval. If NULL, no error bars are created.}
}
\value{
A data frame which can be plotted with lineChartDf. See the documentation for lineChartDf for examples.
}
\description{
This function creates a data frame that can be used for plotting
from a user-supplied data frame. The user data will be aggregated
based on the formula. The resulting plotting data frame can be used
with lineChartDf.
}

