% Generated by roxygen2 (4.0.2): do not edit by hand
\name{drawConnectedPointsDf}
\alias{drawConnectedPointsDf}
\title{Draw points with connect lines.}
\usage{
drawConnectedPointsDf(plotDf)
}
\arguments{
\item{plotDf}{The data frame containing the data to be plotted. it must have several columns with specific names. See the return value of `createPlottingDf` for those column names.}
}
\description{
Draw points specified by a plotting data frame with connect lines. This function is primarily for internal use within the LineChart package. If you want to add connected points to a plot, it will probably be easier to use lineChart with add = TRUE.
}

