% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LineChart-functions.R
\name{extractGroupSettings}
\alias{extractGroupSettings}
\title{Extract the appearance settings from a plotting data frame}
\usage{
extractGroupSettings(plotDf)
}
\arguments{
\item{plotDf}{The plotting data frame to extract settings from.}
}
\value{
The extracted settings.
}
\description{
Extract the appearance settings from a plotting data frame
}
\examples{
data(ChickWeight)
plotDf = lineChart(weight ~ Time * Diet, ChickWeight, legendPosition="topleft")

#From a data frame used for plotting, extract the settings
settings = extractGroupSettings(plotDf)

settings$symbol = 21:24 #Make a small adjustment to the settings

#Re-plot the data with new settings
lineChart(weight ~ Time * Diet, ChickWeight, legendPosition="topleft", settings=settings)
}
