% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LineChart-functions.R
\name{legendFromPlottingDf}
\alias{legendFromPlottingDf}
\title{Create a legend based on a plotting data frame with group appearance settings}
\usage{
legendFromPlottingDf(position, plotDf, ...)
}
\arguments{
\item{position}{The position of the legend. If "CHOOSE_BEST", the best legend position is selected. Otherwise, this is passed to the `x` argument of legend().}

\item{plotDf}{The plotting data frame from which settings will be extracted and used to create the legend.}

\item{...}{A few variable arguments are passed through to legend(): y, cex, inset, title, box.lwd, and horiz.}
}
\description{
Create a legend based on a plotting data frame with group appearance settings
}

