% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LineChart-functions.R
\name{legendFromSettings}
\alias{legendFromSettings}
\title{Create a legend based on group appearance settings}
\usage{
legendFromSettings(position, settings, ...)
}
\arguments{
\item{position}{The position of the legend. This is passed to the `x` argument of legend().}

\item{settings}{The settings data frame that will be used to create the legend.}

\item{...}{A few variable arguments are passed through to legend(): y, cex, inset, title, box.lwd, and horiz.}
}
\description{
Create a legend based on group appearance settings
}
