% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LineChart-functions.R
\name{lineChartManual}
\alias{lineChartManual}
\title{Open the introductory manual for the LineChart package.}
\usage{
lineChartManual()
}
\description{
This function open the introductory manual/vignettes.
}

