% Generated by roxygen2 (4.0.1): do not edit by hand
\name{plotLineChart}
\alias{plotLineChart}
\title{Plot data with a line chart}
\usage{
plotLineChart(formula, data, legendPosition = "topright", settings = NULL,
  errBarType = "SE", title = "", xlab = NULL, ylab = NULL,
  legendTitle = "GROUP_NAME", xlim = NULL, ylim = NULL,
  plotXAxis = TRUE, plotYAxis = TRUE, lwd.axes = par()$lwd, add = FALSE)
}
\arguments{
\item{formula}{A formula of the form "y ~ x * group", where "y", "x", and "group" are names of columns in `data`.}

\item{data}{A data frame containing the data to be plotted.}

\item{legendPosition}{The position at which the legend should be placed. If this is not NULL, it is passed through to legend directly.}

\item{settings}{A data frame containing the appearance settings, such as returned by buildGroupSettings.}

\item{errBarType}{The type of error bars to use. Can be one of "SE", "SD", or "95\%CI" for standard error, standard deviation, or 95\% confidence interval, respectively.}

\item{title}{The title to place at the top of the plot.}

\item{xlab}{The label to place on the x-axis. If NULL (default), the name of the source of data for the x-variable will be used.}

\item{ylab}{The label to place on the y-axis. If NULL (default), the name of the source of data for the y-variable will be used.}

\item{legendTitle}{The title to be used in the legend box. If "GROUP_NAME" (default), the name of the source of data for the grouping variable will be used. If no legend title is desired, use NULL.}

\item{xlim}{The plotting range of the x-axis.}

\item{ylim}{The plotting range of the y-axis.}

\item{plotXAxis}{Should the x-axis be plotted?}

\item{plotYAxis}{Should the y-axis be plotted?}

\item{lwd.axes}{The line width of the axes and box around the plotting area.}

\item{add}{If FALSE, a new plot will be started and the data plotted in it. If TRUE, the data will be plotted in the current plotting device, if any.}
}
\value{
The plotting data frame that was used is returned invisibly.
}
\description{
This function is the primary interface to the LineChart package.
}
\examples{
#Basic use case: Plotting data frame with default appearance.
data(ChickWeight)
plotLineChart(weight ~ Time * Diet, ChickWeight, legendPosition="topleft")

#You can modify the appearance of groups by providing group settings:
settings = buildGroupSettings(group=1:4, symbol=21:24, lty=1:4)
plotLineChart(weight ~ Time * Diet, ChickWeight, legendPosition="topleft", settings=settings)

#Let's ignore group 3
settings[ settings$group == 3, "include" ] = FALSE
plotLineChart(weight ~ Time * Diet, ChickWeight, legendPosition="topleft", settings=settings)

#If for whatever reason you have data in two different data frames that you
#want in the same plot, use add=TRUE to add to the existing plot.
cw12 = ChickWeight[ ChickWeight$Diet \%in\% 1:2, ]
cw34 = ChickWeight[ ChickWeight$Diet \%in\% 3:4, ]

settings$fillColor = c("white", "white", "black", "black") #Distinguish between the groups by fill color.
settings$include = TRUE #Include all groups

plotLineChart(weight ~ Time * Diet, cw12, legendPosition="topleft", settings=settings, ylim=c(40,300))
plotLineChart(weight ~ Time * Diet, cw34, legendPosition="bottomright", settings=settings, add=TRUE)
}

