% Generated by roxygen2 (4.0.2): do not edit by hand
\name{applySettingsToPlottingDf}
\alias{applySettingsToPlottingDf}
\title{Apply appearance settings to a plotting data frame.}
\usage{
applySettingsToPlottingDf(settings, plotDf)
}
\arguments{
\item{settings}{The appearance settings to apply.}

\item{plotDf}{The plotting data frame to apply the settings to.}
}
\value{
A new plotting data frame with settings applied.
}
\description{
Apply appearance settings to a plotting data frame.
}

