% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LineChart-functions.R
\name{createPlottingDf}
\alias{createPlottingDf}
\title{Creates a plottable data frame.}
\usage{
createPlottingDf(formula, data, settings = NULL,
  centralTendencyType = "mean", errBarType = "SE", xOrder = NULL)
}
\arguments{
\item{formula}{The formula should be of the form \code{y ~ x} or \code{y ~ x * g1 * g2 * ... * gK}, where the \code{gN} are any number of grouping variables and \code{y}, \code{x}, and \code{gi} are the names of columns in \code{data}. If grouping variables are used, \code{x} must come before them.}

\item{data}{A data frame containing the data to be used in creation of the plotting data frame.}

\item{settings}{Plotting settings for the different groups in the data, such as the symbol to use.}

\item{centralTendencyType}{Character or function. The type of central tendency measure to use. Can be "mean" or "median". If a function, should be a function of one vector argument that returns a scalar.}

\item{errBarType}{The type of error bar to use. Can be "SE" for standard error, "SD" for standard deviation, or "CI95" for a 95\% confidence interval. If NULL, no error bars are created. If a function is supplied, the function should take one vector argument, which is the data used to plot a single data point. It should return either 1) a length-2 vector or 2) a list with two elements. If returning 1), the values in the vector should be distances from the central tendency measure that the error bars should be drawn (i.e. they should be more-or-less centered on 0). If returning 2), the list should contain \code{eb}, which is either a) error bar distances OR b) error bar endpoints, and \code{includesCenter}, which indicates whether \code{eb} is distances or endpoints. If \code{eb} is distances, then it does not include the center, so \code{includesCenter} should be \code{FALSE}. If \code{eb} is endpoints, then it does include the center, so \code{includesCenter} should be \code{TRUE}.}

\item{xOrder}{The order in which to plot the x variable. A character vector containing all of the levels of the x variable in the order in which they should be plotted. If the \code{x} variable is numeric or can be coerced to numeric, this argument does nothing.}
}
\value{
A data frame which can be plotted with \code{\link{lineChartDf}}. See the documentation for lineChartDf for examples.
}
\description{
This function creates a data frame that can be used for plotting
from a user-supplied data frame. The user data will be aggregated
based on the formula. The resulting plotting data frame can be used
with \code{lineChartDf}.
}
\details{
Note that if no grouping variables are supplied by the formula, a \code{group} column
will still be made in the resulting data frame and the group name will be the number 0.
}
\examples{

data(ChickWeight)

#Example of a user-defined central tendency function:
trimmedMedian = function(x) {
  #trim highest and lowest values
  x = sort(x)[ 2:(length(x) - 1) ]
  median(x)
}

#Example of a user-defined error bar function:
quartiles = function(x) {
  qs = as.numeric(quantile(x, c(0.25, 0.75)))
  list(eb=qs, includesCenter=TRUE)
}

plotDf = createPlottingDf(weight ~ Time * Diet, ChickWeight, 
  centralTendencyType=trimmedMedian, errBarType=quartiles)
lineChartDf(plotDf)

}
