% Generated by roxygen2 (4.0.2): do not edit by hand
\name{legendFromPlottingDf}
\alias{legendFromPlottingDf}
\title{Create a legend based on a plotting data frame with group appearance settings}
\usage{
legendFromPlottingDf(position, plotDf, ...)
}
\arguments{
\item{position}{The position of the legend. This is passed to the `x` argument of legend().}

\item{plotDf}{The plotting data frame from which settings will be extracted and used to create the legend.}

\item{...}{A few variable arguments are passed through to legend(): y, cex, inset, title, box.lwd, and horiz.}
}
\description{
Create a legend based on a plotting data frame with group appearance settings
}

