% Generated by roxygen2 (4.0.2): do not edit by hand
\name{lineChart}
\alias{lineChart}
\title{Plot data with a line chart}
\usage{
lineChart(formula, data, legendPosition = "topright", settings = NULL,
  errBarType = "SE", title = "", xlab = NULL, ylab = NULL,
  legendTitle = "GROUP_NAME", xlim = NULL, ylim = NULL,
  plotXAxis = TRUE, plotYAxis = TRUE, lwd.axes = par()$lwd, add = FALSE,
  ...)
}
\arguments{
\item{formula}{The formula should be of the form `y ~ x` or `y ~ x * g1 * g2 * ... * gK`, where the `gN` are any number of grouping variables and 'y', 'x', and 'gi' are the names of columns in `data`. If grouping variables are used, `x` must come before them.}

\item{data}{A data frame containing the data to be used in creation of the plotting data frame.}

\item{legendPosition}{The position at which the legend should be placed. If this is not NULL, it is passed through to legend directly.}

\item{settings}{Plotting settings for the different groups in the data, such as the symbol to use.}

\item{errBarType}{The type of error bar to use. Can be "SE" for standard error, "SD" for standard deviation, or "CI95" for a 95\% confidence interval. If NULL, no error bars are created. If a function is supplied, the function should take one argument, which will be the data used to plot a single data point, and return two error bar amounts. These amounts should be the signed difference between the end of the error bar and the mean of the data for that data point.}

\item{title}{The title to place at the top of the plot.}

\item{xlab}{The label to place on the x-axis. If NULL (default), the name of the source of data for the x-variable will be used, if available.}

\item{ylab}{The label to place on the y-axis. If NULL (default), the name of the source of data for the y-variable will be used if available.}

\item{legendTitle}{The title to be used in the legend box. If "GROUP_NAME" (default), the name of the source of data for the grouping variable will be used. If no legend title is desired, use NULL.}

\item{xlim}{The plotting range of the x-axis.}

\item{ylim}{The plotting range of the y-axis.}

\item{plotXAxis}{Should the x-axis be plotted?}

\item{plotYAxis}{Should the y-axis be plotted?}

\item{lwd.axes}{The line width of the axes and box around the plotting area.}

\item{add}{If FALSE, a new plot will be started and the data plotted in it. If TRUE, the data will be plotted in the current plotting device, if any.}

\item{...}{Additional parameters, currently passed on to the legend creating functions.}
}
\value{
The plotting data frame that was used is returned invisibly.
}
\description{
This function is the primary interface to the LineChart package.
}
\examples{
#Basic use case: Plotting data frame with default appearance.
data(ChickWeight)
lineChart(weight ~ Time * Diet, ChickWeight, legendPosition="topleft")

#You can modify the appearance of groups by providing group settings:
settings = buildGroupSettings(group=1:4, symbol=21:24, lty=1:4)
lineChart(weight ~ Time * Diet, ChickWeight, legendPosition="topleft", settings=settings)

#Let's ignore group 3
settings[ settings$group == 3, "include" ] = FALSE
lineChart(weight ~ Time * Diet, ChickWeight, legendPosition="topleft", settings=settings)

#If for whatever reason you have data in two different data frames that you
#want in the same plot, use add=TRUE to add to the existing plot.
cw12 = ChickWeight[ ChickWeight$Diet \%in\% 1:2, ]
cw34 = ChickWeight[ ChickWeight$Diet \%in\% 3:4, ]

#Distinguish between the groups by fill color.
settings$fillColor = c("white", "white", "black", "black")
settings$include = TRUE #Include all groups

lineChart(weight ~ Time * Diet, cw12, legendPosition="topleft",
  settings=settings, ylim=c(40,300))
lineChart(weight ~ Time * Diet, cw34, legendPosition="bottomright",
  settings=settings, add=TRUE)
}
\seealso{
\code{\link{createPlottingDf}}, \code{\link{lineChartDf}}, \code{\link{buildGroupSettings}}, \code{\link{extractGroupSettings}}, \code{\link{applySettingsToPlottingDf}}, \code{\link{legendFromPlottingDf}}, \code{\link{legendFromSettings}}
}

